runtime: split out `Parser` into a `parser` module.
