+++ src/lib.rs
@@ -16,6 +16,8 @@ pub mod input;
 #[forbid(unsafe_code)]
 pub mod parse_node;
 #[forbid(unsafe_code)]
+pub mod parser;
+#[forbid(unsafe_code)]
 pub mod proc_macro;
 #[forbid(unsafe_code)]
 pub mod runtime;
